refactor: centralize option dataclasses
